gui: Add awaitable restart and exit functions
  Electron's `app.exit()` function is asynchronous but does not return a
  Promise so the process calling it will execute code coming after while
  it's still running.

  To avoid this, we add 2 functions built around `app.exit()` to exit or
  restart the Desktop client and be quite sure no code coming after
  these asynchronous calls will be executed.
  We simply wait for a second after calling `app.exit()`. This should be
  enough for the process to fully exit.
